+++ lib/anyway_config.rb
@@ -10,6 +10,7 @@ require "anyway/version"
 require "anyway/ext/deep_dup"
 require "anyway/ext/deep_freeze"
 require "anyway/ext/hash"
+require "anyway/ext/flatten_names"
 
 require "anyway/utils/deep_merge"
 require "anyway/utils/which"